requirements.txt
praw
openai
python-dotenv
langchain
tqdm
rich
